Using doubles instead for intermediate points of gravity

+++ modules/GravityPlugin/src/main/java/org/gephi/plugins/layout/forceAtlas2Custom/ForceAtlas2LayoutData.java
@@ -56,5 +56,5 @@ public class ForceAtlas2LayoutData implements LayoutData {
     public double old_dy = 0;
     public double mass = 1;
     // Gravity source
-    public int gravitySource = 0;
+    public double gravitySource = 0;
 }
